+++ src/main/java/com/structurizr/dsl/DeploymentEnvironment.java
@@ -29,7 +29,7 @@ class DeploymentEnvironment extends Element {
     }
 
     @Override
-    protected Set<String> getRequiredTags() {
+    public Set<String> getDefaultTags() {
         return null;
     }