+++ src/main/java/com/structurizr/dsl/ElementGroup.java
@@ -36,7 +36,7 @@ class ElementGroup extends Element {
     }
 
     @Override
-    protected Set<String> getRequiredTags() {
+    public Set<String> getDefaultTags() {
         return null;
     }