+++ src/main/java/com/structurizr/dsl/ModelItemParser.java
@@ -10,9 +10,6 @@ final class ModelItemParser extends AbstractParser {
 
     private final static int URL_INDEX = 1;
 
-    private final static int PROPERTY_NAME_INDEX = 0;
-    private final static int PROPERTY_VALUE_INDEX = 1;
-
     private final static int PERSPECTIVE_NAME_INDEX = 0;
     private final static int PERSPECTIVE_DESCRIPTION_INDEX = 1;
 
@@ -54,23 +51,6 @@ final class ModelItemParser extends AbstractParser {
 
         String url = tokens.get(URL_INDEX);
         context.getModelItem().setUrl(url);
-    }
-
-    void parseProperty(ModelItemPropertiesDslContext context, Tokens tokens) {
-        // <name> <value>
-
-        if (tokens.hasMoreThan(PROPERTY_VALUE_INDEX)) {
-            throw new RuntimeException("Too many tokens, expected: <name> <value>");
-        }
-
-        if (tokens.size() != 2) {
-            throw new RuntimeException("Expected: <name> <value>");
-        }
-
-        String name = tokens.get(PROPERTY_NAME_INDEX);
-        String value = tokens.get(PROPERTY_VALUE_INDEX);
-
-        context.getModelItem().addProperty(name, value);
     }
 
     void parsePerspective(ModelItemPerspectivesDslContext context, Tokens tokens) {