+++ src/main/java/com/structurizr/dsl/PropertiesDslContext.java
@@ -1,0 +1,17 @@
+package com.structurizr.dsl;
+
+import com.structurizr.PropertyHolder;
+
+final class PropertiesDslContext extends DslContext {
+
+    private PropertyHolder propertyHolder;
+
+    public PropertiesDslContext(PropertyHolder propertyHolder) {
+        this.propertyHolder = propertyHolder;
+    }
+
+    PropertyHolder getPropertyHolder() {
+        return this.propertyHolder;
+    }
+
+}